refactor(type): move nullable out of `DataType` (#836)
This PR moves `nullable` out of `DataType` and puts it into
`ColumnDesc`. The original `DataTypeKind` is renamed to `DataType`.

This is to make it align with other projects like arrow-rs and
RisingWave. Actually I'm planning to migrate the entire array module to
arrow-rs, so that we can be more standard and have better
interoperability with the community.

---------

Signed-off-by: Runji Wang <wangrunji0408@163.com>