feat: support WITH clause (#838)
This PR adds support for specifying CTEs using the `WITH` clause. We
treat CTEs just as named subqueries. The implementation simply adds CTE
queries to the binder context and inlines them when referenced.

However, due to an optimizer bug (which has been fixed in #796), the
current CTE is pretty useless as most practice use cases will lead to
panics.

---------

Signed-off-by: Runji Wang <wangrunji0408@163.com>