feat: support view in memory (#837)
This PR is a part of #796, adds support for creating, querying and
dropping views in memory.

The key implementations are: 
1. When creating a view, bind the query and store the logical plan with
the view in catalog.
2. When querying from a view, build executors for all views and then
build other plan nodes on top of them. Given that a view can be consumed
by multiple downstream nodes, we introduce `StreamSubscriber` to allow
multiple consumers of a stream.

Limitations:
1. We don't persist views in disk storage.
2. We don't support inferring schema from the query. Columns must be
defined explicitly when creating a view.
3. We don't maintain dependency relationship between tables and views.

---------

Signed-off-by: Runji Wang <wangrunji0408@163.com>